Configured rollup with Vue 3 + reconfigured rollup settings
