+++ lib/package.json
@@ -1,7 +1,7 @@
 {
   "name": "vue-easy-dnd",
-  "version": "2.0.0",
-  "description": "Easy-DnD is a drag and drop implementation for Vue.js 3 that uses only standard mouse events instead of the HTML5 drag and drop API, which is [impossible to work with](https://www.quirksmode.org/blog/archives/2009/09/the_html5_drag.html). Think of it as a way to transfer data from some components to others using the mouse or support for a mouse assisted copy/cut - paste. It also allows for lists to be reordred by drag and drop.",
+  "version": "2.0.0-beta.1",
+  "description": "Easy-DnD is a drag and drop implementation for Vue 3 that uses only standard mouse events instead of the HTML5 drag and drop API, which is [impossible to work with](https://www.quirksmode.org/blog/archives/2009/09/the_html5_drag.html). Think of it as a way to transfer data from some components to others using the mouse or support for a mouse assisted copy/cut - paste. It also allows for lists to be reordred by drag and drop.",
   "main": "./dist/vue-easy-dnd.ssr.js",
   "module": "./dist/vue-easy-dnd.esm.js",
   "browser": {
@@ -28,6 +28,8 @@
   "keywords": [
     "vue",
     "vuejs",
+    "vue2",
+    "vuejs2",
     "vue3",
     "vuejs3",
     "drag-and-drop",
@@ -39,5 +41,8 @@
   "bugs": {
     "url": "https://github.com/rlemaigre/Easy-DnD/issues"
   },
-  "homepage": "https://github.com/rlemaigre/Easy-DnD#readme"
+  "homepage": "https://github.com/rlemaigre/Easy-DnD#readme",
+  "engines": {
+    "node": ">= 14"
+  }
 }
