+++ lib/src/index.js
@@ -1,12 +1,15 @@
 import Drag from './components/Drag.vue';
 import Drop from './components/Drop.vue';
 import DropMask from './components/DropMask.vue';
+import DropList from './components/DropList.vue';
+import DragFeedback from './components/DragFeedback.vue';
+
 import DragAwareMixin from './mixins/DragAwareMixin';
 import DragMixin from './mixins/DragMixin';
 import DropMixin from './mixins/DropMixin';
+
 import { dnd } from './js/DnD';
 import { DragImagesManager } from './js/DragImagesManager';
-import DropList from './components/DropList.vue';
 import { DnDEvent, InsertEvent, ReorderEvent } from './js/events';
 import { createDragImage } from './js/createDragImage';
 
@@ -15,6 +18,7 @@ export {
   Drop,
   DropList,
   DropMask,
+  DragFeedback,
   DragAwareMixin,
   DragMixin,
   DropMixin,
