+++ package.json
@@ -4,8 +4,8 @@
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
-    "build-dev": "rollup --config ./lib/rollup.config.js",
-    "build": "rollup --config ./lib/rollup.config.js --environment BUILD:production",
+    "build-dev": "cd ./lib && rollup --config rollup.config.mjs",
+    "build": "cd ./lib && rollup --config rollup.config.mjs --environment BUILD:production",
     "lint": "vue-cli-service lint ./src ./lib"
   },
   "dependencies": {
@@ -18,18 +18,17 @@
   "devDependencies": {
     "@babel/core": "^7.12.16",
     "@babel/eslint-parser": "^7.12.16",
+    "@rollup/plugin-node-resolve": "^15.0.1",
     "@vue/cli-plugin-babel": "~5.0.0",
     "@vue/cli-plugin-eslint": "~5.0.0",
     "@vue/cli-service": "~5.0.0",
     "eslint": "^7.32.0",
     "eslint-plugin-css": "^0.6.0",
     "eslint-plugin-vue": "^8.0.3",
-    "rollup": "^1.20.3",
-    "rollup-plugin-buble": "^0.19.8",
-    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup": "^3.7.4",
     "rollup-plugin-cpy": "^2.0.1",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-sass": "^1.12.17",
+    "rollup-plugin-delete": "^2.0.0",
+    "rollup-plugin-scss": "^4.0.0",
     "rollup-plugin-vue": "^6.0.0",
     "sass": "^1.56.1",
     "sass-loader": "^13.2.0"
@@ -77,8 +76,14 @@
       "vue/require-explicit-emits": "off",
       "space-before-function-paren": "error",
       "prefer-const": "error",
-      "object-curly-spacing": ["error", "always"],
-      "quotes":["error", "single"],
+      "object-curly-spacing": [
+        "error",
+        "always"
+      ],
+      "quotes": [
+        "error",
+        "single"
+      ],
       "indent": [
         "error",
         2
